feat: refatoramento para padrao mvc
